PO进展中

+++ CMASS/src/edu/nju/dataservice/collectiondataservice/CollectionDataService.java
@@ -7,7 +7,7 @@ import edu.nju.po.*;
 
 public interface CollectionDataService {
 	public void saveSendDocPO(SendDocPO po);
-	public int getDistance(String city);
+	public Double getDistance(String city);
 	public int getSequence();
 	public void changeSequence(int sequence);
 	public void saveHistory(HistoryPO po);
